Refactored the application. Changed project names. Created project based on clean architecture.
