+++ Domain/Enitities/Activity.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain
+namespace Domain.Enitities
 {
     public class Activity
     {
