+++ Reactivities.sln
@@ -5,9 +5,13 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API", "API\API.csproj", "{CA78704E-E264-49A2-997C-DDDDD2994BD2}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAccess", "Persistence\DataAccess.csproj", "{3D11ABCE-012D-4886-BFEE-6A90870B5326}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Persistence", "Persistence\Persistence.csproj", "{3D11ABCE-012D-4886-BFEE-6A90870B5326}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Server", "Server\Server.csproj", "{5DF9A9C9-B975-4620-9632-0DB72CFCCE30}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthentificationServer", "Server\AuthentificationServer.csproj", "{5DF9A9C9-B975-4620-9632-0DB72CFCCE30}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Application\Application.csproj", "{A677454F-6D10-43D6-ADA8-9D4EB3B1DE5C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "Domain\Domain.csproj", "{EC102804-AF8F-4A62-B409-7C27F71B5B6B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +31,14 @@ Global
 		{5DF9A9C9-B975-4620-9632-0DB72CFCCE30}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5DF9A9C9-B975-4620-9632-0DB72CFCCE30}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5DF9A9C9-B975-4620-9632-0DB72CFCCE30}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A677454F-6D10-43D6-ADA8-9D4EB3B1DE5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A677454F-6D10-43D6-ADA8-9D4EB3B1DE5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A677454F-6D10-43D6-ADA8-9D4EB3B1DE5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A677454F-6D10-43D6-ADA8-9D4EB3B1DE5C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC102804-AF8F-4A62-B409-7C27F71B5B6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC102804-AF8F-4A62-B409-7C27F71B5B6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC102804-AF8F-4A62-B409-7C27F71B5B6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC102804-AF8F-4A62-B409-7C27F71B5B6B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
